Trapdoors!
- Moved more block-related functionality from helper to blocks
- Added IsInteractable and InteractWithBlock
- Trapdoors are now fully functional!
