+++ src/misc/helpers/helper.h
@@ -15,13 +15,11 @@
 #include "logger.h"
 #include "blocks.h"
 #include "items.h"
-#include "directions.h"
 #include "datatypes.h"
 
 Int3 LocalToGlobalPosition(Int3 chunkPos, Int3 blockPos);
 Int3 BlockToChunkPosition(Vec3 position);
 Int3 BlockToChunkPosition(Int3 position);
-void BlockToFace(int32_t& x, int8_t& y, int32_t& z, int8_t& direction);
 
 int8_t ConvertFloatToPackedByte(float value);
 Vec3 SubtractVec3(Vec3 previousPosition, Vec3 currentPosition);
@@ -44,8 +42,6 @@ Int3 DecodeChunkHash(int64_t hash);
 
 int32_t SafeStringToInt(std::string in);
 int64_t SafeStringToLong(std::string in);
-
-Block GetPlacedBlock(int32_t x, int8_t y, int32_t z, int8_t face, int8_t playerDirection, int16_t id, int16_t damage);
 
 std::string GetRealTime();
 std::string GetRealTimeFileFormat();
